tools/requirements.txt
polars==1.17.1
python-dotenv==1.0.1